build: add some workarounds for Windows, explicitly link dependencies
CMake prior to 3.16 did not correctly emit the import library into the requested
location.  Adjust the library emission path instead to workaround missing import
libraries.  Explicitly link against Foundation in TSCBasic which is needed on
non-Darwin targets.  Add the ability to build against an explicit build of
Foundation or the toolchain provided one.