Basic: handle root checking on Windows more properly
Roots on Windows are more complicated.  The drive name is not considered
a root, but is a root for the Path API as you can not go any further.
It is however a name and not a path, therefore it is not a root path.
Use that alongwith `PathCchIsRootW` to classify the paths on Windows as
being a root or not.