TSCBasic: use the "portable" names for `terminalType`
This adds the Windows path which uses the `_isatty` and `_fileno` for
the conversion.  This is likely still insufficiently to properly detect
the stream.  `_isatty` is okay in the short term, but it is better to
use `GetFileType` on the `HANDLE` associated with the file descriptor.
This requires better auditing to ensure that the handle is convertible
to the underlying Windows handle.