build: make `SQLite` a private dependency
Since the SQLite3 import is marked as `@_implementationOnly` this is
perfectly reasonable to assume that the symbols from SQLite3 should not
be required by consumers of this library and thus the consumers of this
library do not need to link against SQLite3 themselves.  Reflect this in
the build.  This is needed to ensure that the dependency handling works
properly for building swift-driver.