build: synchronise CMake and Package.swift for Windows
Add the additional link dependency on Windows and a macro definition to
avoid some spew from the library build.  Note that this is using an OS
specific block to alter the settings rather than the inline definition
since older versions of s-p-m may crash on the application of the
setting.

This requires bumping the tools version to 5.2.