TSCBasic: correct `resolveSymlink` on Windows
This would fail to properly resolve the symbolic links on Windows since
the result of the resolution would be a path relative to the location of
the original path.  However, we would instead treat the path as the
absolute path.  Correct this which is required to resolve modulemaps
inside of directories with symbolic links.

This was found by building IndexStoreDB on Windows with SwiftPM.