[NFC] Switch BitstreamReader Entrypoints to Use ByteString
ByteString is a more appropriate vocabulary type than Data to use for
these entrypoints. TSCBasic traffics in ByteString values, and its
representation as an array means that there are reasonable indexing
invariants we can rely on with respect to subsequences.

This should also enable some clean up in the swift-driver.