Process: only close the read end of the pipe after reading the full content successfully
If we close the read end of the pipe after hitting an error (EIO for example),
the child process may be still running and will print more content into the write end of the
pipe. Such write to a closed-read pipe will cause a SIGPIPE signal and fail.

Potentiall a fix to rdar://74058113