Basic OpenBSD support.
Of note is the changes to FSWatch. inotify is only available on Linux,
so do not condition on `canImport(Glibc)` but restrict to operating
systems using the Linux kernel only.

Naturally, OpenBSD does not have FSEventStream either, so for now,
create a no-op delegate for this platform. It is possible that this
particular functionality might be provided by some creative use of
kevent, but we defer that here: such behavior would be important for
platforms that do not have support for monitoring filesystem events.

On this platform, SQLite resides in /usr/local/lib, so make the
necessary cmake changes as well, while we are here.

We do still need to do something about the `unversionedTriple`
inference: this will be handled separately.