Respect `SWIFT_TOOLCHAIN_MACOS_DEPLOYMENT_TARGET` when set (such as during Swift toolchain build) (#220)
If it is not set in the environment, stay with the default minimum deployment target of 10.10 on macOS.

This aligns with both swift-package-manager and with swift-driver, and is important now that ToolSupportCore has started containing some code that is annotated with availability for Darwin platforms.  It uses the more project-neutral name `SWIFT_TOOLCHAIN_MACOS_DEPLOYMENT_TARGET` and the expectation is that both SwiftPM and SwiftDriver will switch to using it as well.