Use the (versioned) triple from swiftc instead.
Here, swiftc is invoked to get the textual unversioned triple from the
compiler. However, TSC wants to use this unversioned triple for the
target flag, which causes breakage on the compiler side on platforms
with versioned triples, because the compiler can't find the stdlib with
the unversioned triple.

Instead, default to getting versioned triples. When clients of TSC want
to do funky things with unversioned triples, we lop the version numbers
off versioned triples to get an unversioned triple. This is somewhat
crude, but the alternative is to get both versioned _and_ unversioned
triples from swiftc, which is similarly awkward.