Switch to an Unowned View of Bitstream Record Data
Now that we no longer have any entrypoints that can escape this data, switch to a non-owning view of bitstream data. This saves a significant amount of copying overhead when deserializing many/large bitstream files and allows callers to decide for themselves what format they wish to reinterpret the data as.