build: disable in-tree builds
This ensures that one does not accidentally build t-s-c in tree, which
means that the source tree is always pristine and the build tree root is
different from the source tree.  The previous change to ignore CMake
artifacts would only be caused by an accidental in-tree build which
should be frowned upon and be treated as unsupported.