TSCBasic: rework relative path construction on Windows (#298)
When constructing a path relative to a root, compose the final path
before canonicalization as the canonicalization may require lookups
in the base location.

We additionally pay a small penalty to convert the slashes to native
separators.  This is purely for aesthetic purposes, it ensures that the
path is computed initially rather than delayed until use.

This in conjunction with apple/swift-package-manager#4237 enables the
use of header search paths in SPM which are based in other targets.