TSCBasic: change spelling for `root`
Unlike Unix, Windows does not have a singular unified file system root.
Instead, you have a forest with up to 26 roots (A-Z).  Due to the
current design limitations of the Path API, use a drive-relative path as
a root to terminate recursion when using the `InMemoryFileSystem` which
is required to work towards enabling tests for SPM.