TSCTestSupport: support more complex file systems
File systems are not guaranteed to be uniformly accessible through a
singular view.  Always use the current working directory to determine
the of the temporary directory as the path may not be viable for
conversion to a relative path.  This is mandatory for resolving
assumptions in the swift-package-manager test suite.