TSCBasic: alter environment handling for Windows
Rather than using the unicode environment, which should be preferred,
use the ANSI environment operations.  This is important to do here as
the reading of the environment relies on Foundation, which in turn
relies on CoreFoundation for `_CFEnviron` which returns the ANSI
environment.  As a result, we need to ensure that we use the ANSI
environment variables through out as the C runtime will maintain the two
environments in parallel and may go out of sync.

This will enable additional tests in Swift Package Manager to pass on
Windows, but should really be addressed by changing Foundation to use
the Unicode environment variables and then using that once more here.