Update maps in the reader
This commit does some additional cleanup. Instead of iterating over the
records twice while constructing the diagnostic itself, the reader can
load everything into the map while reading. This removes the extra loop
over the records, and means that the diagnostic constructor doesn't need
to mutate the state of the maps.