TSCBasic: normalise the drive letter spelling on Windows
Windows normalises the drive letter to an uppercase letter.  However,
some environments (e.g. nodejs) use the lower case spelling.  This
results in a path spelling difference which triggers a rebuild in the
case that the LSP is running within VSCode.

Normalise the spelling to the upper case always when building an
`AbsolutePath`.

Based on a patch by Ami Fischman!

Co-authored-by: Ami Fischman <ami@thebrowser.company>