TSCBasic: avoid using C library functions for environment
Prefer to use the Win32 APIs for environment management.  Note that this
is going to break any user of libc on Windows.  Setting environment
variables through this function will not be reflected in the C library
only the Win32 APIs.  Furthermore, we use the unicode variants always as
the unicode and ANSI environment may diverge as not all unicode (UTF-16)
is translatable to ANSI, which is documented at [1].

[1] https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/getenv-wgetenv?view=msvc-170&viewFallbackFrom=vs-2019.