Add Swift Testing helper functions
In the process of converted XCTests to Swift Testing, add a few Swift
Testing helper functions equivalent to the XCTest equivalents.

Unfortunately, I'm unable to write automated tests that verify the new
helpers behave as expected as the Swift Testing framework does not
currently allow writing tests that verify an issue was recorded. This is
tracked under https://github.com/swiftlang/swift-testing/issues/155